requirements.txt
streamlit
tensorflow-lite
pandas
Pillow
plotly